Weather: Allow syncing weather based on actual location
* Ask for FINE Location in PermissionsActivity
* Add user position on the Weather Map
* OptIn: Allow automatically getting weather based on user Location

If this option is enabled, the half hourly sync will not use the userdefined location but instead get the Location via GPS/Network.
This can be enabled in the Weather settings dialog.
If a user decides not to use this feature the GPS will only be used while on the Position Picker to show the current position.