suppress QUERY_ALL_PACKAGES lint
The use of QUERY_ALL_PACKAGES is not allowed on GooglePlay and the
linter errors out with a notice that specific queries should be used.

Queries can be used to get apps by:
	- Specific package names
	- Packages that match an intent filter signature
	- Packages that use a specific authority

This does not cover our scenario of dynamic package information
retrieval, that we need for the notifications.

Google has acknowledged the requirement of `QUERY_ALL_PACKAGES` for
"Device management apps". [1]

[1]: https://developer.android.com/training/package-visibility/declaring#all-apps